webhook-alert-v2/webhook-alert/requirements.txt
requests==2.26.0
flask==2.0.2
urllib3==1.26.7
jinja2==3.0.3
pyyaml==6.0
